+++ cast/__init__.py
@@ -1,7 +1,25 @@
-__version__ = "0.1.8"
+__version__ = "0.1.12"
+from pathlib import Path
+
+
+def podcast_same_audio_case(audio_model, file_name, context):
+    """
+    podcast special case:
+
+    Upload all audio fiels with the same name into same
+    model instance, because podcast audio consists of
+    different files with the same content.
+    """
+    stemmed_filename = Path(file_name).stem
+    for audio in audio_model.objects.order_by("-created")[:10]:
+        # FIXME use index instead of limit trick (maybe)
+        if stemmed_filename in audio.get_audio_file_names():
+            context["instance"] = audio
+    return context
 
 
 def upload_handler(request):
+    # Apps aren't loaded yet.
     from . import models
     from filepond.forms import get_model_form
 
@@ -9,15 +27,21 @@ def upload_handler(request):
     for ending in ("jpg", "jpeg", "png", "gif"):
         lookup[ending] = (models.Image, "original", "user")
 
-    for ending in ("wav", "webm", "ogg", "mp3", "m4a"):
-        lookup[ending] = (models.Audio, ending, "user")
+    for ending in ("wav", "webm", "ogg", "mp3", "m4a", "opus"):
+        upload_field_name = ending
+        if ending == "ogg":
+            upload_field_name = "oga"
+        lookup[ending] = (models.Audio, upload_field_name, "user")
 
     for ending in ("mp4", "mov", "m4v"):
         lookup[ending] = (models.Video, "original", "user")
 
     file_name = str(request.FILES["original"])
-    ending = file_name.lower().split(".")[-1]
+    ending = Path(file_name).suffix.split(".")[-1].lower()
     local_model, upload_field_name, user_field = lookup[ending]
-    return get_model_form(
+    form_class, context = get_model_form(
         local_model, upload_field_name=upload_field_name, user_field=user_field
     )
+    if local_model.__name__ == "Audio":
+        context = podcast_same_audio_case(models.Audio, file_name, context)
+    return form_class, context