+++ cast/models.py
@@ -6,8 +6,8 @@ import logging
 import tempfile
 import subprocess
 
+from pathlib import Path
 from subprocess import check_output
-
 from collections import defaultdict
 
 from django.db import models
@@ -269,6 +269,25 @@ class Audio(TimeStampedModel):
             field = getattr(self, name)
             if field.name is not None and len(field.name) > 0:
                 yield name, field
+
+    @property
+    def file_formats(self):
+        return " ".join([n for n, f in self.uploaded_audio_files])
+
+    def get_audio_file_names(self):
+        audio_file_names = set()
+        for audio_format, field in self.uploaded_audio_files:
+            audio_file_names.add(Path(field.name).stem)
+        return audio_file_names
+
+    @property
+    def name(self):
+        if self.title is not None:
+            return self.title
+        return ",".join([_ for _ in self.get_audio_file_names()])
+
+    def __str__(self):
+        return f"{self.pk} - {self.name}"
 
     def get_all_paths(self):
         paths = set()